src/agent-memory-layer/requirements.txt
mem0ai
ollama
qdrant-client
python-dotenv
